requirements.txt
flask==2.3.2
jinja2==3.1.2
psutil
psycopg2-binary
flask-login
logging